Refactor to class based structure
