+++ lib/core/state/mongo-server.js
@@ -1,51 +1,20 @@
-const ReplicationCommands = require('../commands/replication-commands')
-
-// Direct mappings to https://docs.mongodb.com/manual/reference/replica-states/
-const SERVER_REPLICATION_STATUS = {
-  UNINITIALIZED: -1, // custom hydra state
-  STARTUP: 0,
-  PRIMARY: 1,
-  SECONDARY: 2,
-  RECOVERING: 3,
-  STARTUP2: 5,
-  UNKNOWN: 6,
-  ARBITER: 7,
-  DOWN: 8,
-  ROLLBACK: 9
-}
+const MongoClient = require('../clients/mongodb-client')
 
 class MongoServerState {
-  constructor(hostname, client) {
-    this.client = client
-    this.hostname = hostname
-    this.replicationStatus = SERVER_REPLICATION_STATUS.UNKNOWN
-  }
-
-  async refreshState() {
-    const replicationCommands = new ReplicationCommands(this.client)
-    let status
-
-    try {
-      status = await replicationCommands.status()
-    } catch (error) {
-      if (!error.code) throw error // Not a MongoError, cannot handle this
-      status = error
-    }
-
-    if (status.code && status.code === 94) {
-      this.replicationStatus = SERVER_REPLICATION_STATUS.UNINITIALIZED
-      return
-    }
-
-    if (status.code) { throw status } // Cannot handle this code
-
-    if (status.myState !== undefined) {
-      this.replicationStatus = status.myState
-      return
-    }
-
-    throw new Error(`Unknown state of replica member: ${JSON.stringify(status)}`)
+  /**
+   *
+   * @param {Object} settings Settings
+   * @param {String} settings.host host used internally by mongo to resolve
+   * @param {String} [settings.uri] host overwrite used by hydra to resolve
+   * @param {*} client MongoClient overwrite for stubs
+   */
+  constructor(settings) {
+    // host
+    this.host = settings.host // Used internally by mongo
+    this.uri = settings.uri || settings.host // Used by hydra
+    // Client
+    this.client = settings.client || new MongoClient(this.uri)
   }
 }
 
-module.exports = { MongoServerState, SERVER_REPLICATION_STATUS }
+module.exports = MongoServerState
