+++ lib/core/state/replication.js
@@ -1,35 +1,32 @@
 const promiseRetry = require('promise-retry')
 const ReplicationCommands = require('../commands/replication-commands')
-const { SERVER_REPLICATION_STATUS } = require('./mongo-server')
+const ReplicaMemberMongoServerState = require('./mongo-server-replica-member')
+const { REPLICA_SET_STATUS, SERVER_REPLICATION_STATUS } = require('../helpers/enums')
 const logger = require('../../helpers/logger')
 
-const REPLICA_SET_STATUS = {
-  UNKNOWN: 0,
-  UNINITIALIZED: 94,
-  HEALTHY: 1
-}
-
 class ReplicaSetState {
-  constructor(configuration) {
-    this.configuration = configuration
+  constructor(settings) {
+    this.replicaSetName = settings.replicaSetName
+    this.servers = settings.servers.map((server) => new ReplicaMemberMongoServerState(server))
+    this.replicaSetStatus = REPLICA_SET_STATUS.UNKNOWN
   }
 
   async waitUntilHealthy() {
     await promiseRetry(async (retry, number) => {
       await this.refreshState()
-      logger.debug(`${this.configuration.replicaSetName}: (${number}) Waiting for state HEALTHY, found "${this.replicaSetStatus}"`)
+      logger.debug(`${this.replicaSetName}: (${number}) Waiting for state HEALTHY, found "${this.replicaSetStatus}"`)
       if (this.replicaSetStatus !== REPLICA_SET_STATUS.HEALTHY) {
-        retry(new Error(`${this.configuration.replicaSetName}: Replica set did not become healthy`))
+        retry(new Error(`${this.replicaSetName}: Replica set did not become healthy`))
       }
     })
   }
 
   async initialize() {
-    logger.debug(`${this.configuration.replicaSetName}: Initializing replica set`)
+    logger.debug(`${this.replicaSetName}: Initializing replica set`)
     // Initiate
-    const { client } = this.configuration.servers[0] // Just pick the first for this
+    const { client } = this.servers[0] // Just pick the first for this
     const replicationCommands = new ReplicationCommands(client)
-    const mongoConfig = this.configuration.toMongoConfiguration()
+    const mongoConfig = this.toMongoConfiguration()
     const initiationStatus = await replicationCommands.initiate(mongoConfig)
     if (initiationStatus.ok !== 1) throw new Error(`Failed replica set initiation with: ${JSON.stringify(initiationStatus)}`) // failure
     await this.waitUntilHealthy() // Wait for election to finish
@@ -37,31 +34,33 @@ class ReplicaSetState {
 
   async apply() {
     await this.refreshState()
-    logger.debug(`${this.configuration.replicaSetName}: Replica set in state "${this.replicaSetStatus}"`)
+    logger.debug(`${this.replicaSetName}: Replica set in state "${this.replicaSetStatus}"`)
 
-    if (this.replicaSetStatus === REPLICA_SET_STATUS.UNINITIALIZED) {
-      await this.initialize()
-      return
+    // All states we can handle
+    const actions = {
+      [REPLICA_SET_STATUS.UNINITIALIZED]: async () => {
+        await this.initialize()
+      },
+      [REPLICA_SET_STATUS.HEALTHY]: async () => {
+        // TODO: Add/remove members, for now we are happy with this though
+      }
     }
 
-    if (this.replicaSetStatus === REPLICA_SET_STATUS.HEALTHY) {
-      return // TODO: Add/remove members, for now we are happy with this though
-    }
-
-    throw new Error(`${this.configuration.replicaSetName}: Replica set status unknown. Cannot apply configuration.`)
+    const action = actions[this.replicaSetStatus]
+    if (!action) throw new Error(`${this.replicaSetName}: Replica set status unknown. Cannot apply configuration.`)
+    return await action()
   }
 
   async refreshState() {
-    logger.debug(`${this.configuration.replicaSetName}: Refreshing status of replica set`)
-    const { servers } = this.configuration
+    logger.debug(`${this.replicaSetName}: Refreshing status of replica set`)
 
     // First connect to all members
-    for (const server of servers) {
-      await server.refreshState()
+    for (const server of this.servers) {
+      await server.updateReplicationState()
     }
 
     // Check if none of the servers are intialized
-    const allUninitialized = servers.every(
+    const allUninitialized = this.servers.every(
       (server) => server.replicationStatus === SERVER_REPLICATION_STATUS.UNINITIALIZED
     )
     if (allUninitialized) {
@@ -70,7 +69,7 @@ class ReplicaSetState {
     }
 
     // Check if all of the servers are either a PRIMARY or SECONDARY
-    const fullyInitialized = servers.every(
+    const fullyInitialized = this.servers.every(
       (server) => server.replicationStatus === SERVER_REPLICATION_STATUS.PRIMARY
         || server.replicationStatus === SERVER_REPLICATION_STATUS.SECONDARY
 
@@ -81,6 +80,20 @@ class ReplicaSetState {
     }
 
     this.replicaSetStatus = REPLICA_SET_STATUS.UNKNOWN
+  }
+
+  /**
+   * Returns the configuration in the structure required by the MongoDB commands:
+   * - replSetInitiate
+   * - replSetReconfig
+   */
+  toMongoConfiguration() {
+    const members = this.servers.map((server, index) => ({ _id: index, host: server.host }))
+    const configuration = {
+      _id: this.replicaSetName,
+      members
+    }
+    return configuration
   }
 }
